+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 ## [0.3.0] - Under development
 
 ### Added
+
 - `"f"`, `"n"`, `"p"` key bindings - Find/Next/Prev (case-insensitive text search)
 - `"l"` key binding - Go to line number
 - `"t"` key binding - Go to timestamp
+- `--timestamp_format` command line option to define a custom timestamp format template,
+  for parsing log files with formats that do not begin with a leading timestamp
 - strip escape sequences from log file lines, for consistent text alignment in columns